requirements.txt
asgiref==3.3.4
Django==3.2
django-redis==5.0.0
Pillow==8.0.0
PyJWT==2.1.0
pytz==2021.1
redis==3.5.3
sqlparse==0.4.1
